Minor refactoring (extract RandomIdGenerator to a separate class)
